✨ feat: 배포 버전 업, 초기화 기능 추가 전체 API 응답 스키마 및 에러 응답 구조 개선 DTO 파일 네이밍 규칙 및 디렉토리 구조 개선

+++ src/users/profiles/profiles.service.ts
@@ -1,5 +1,6 @@
 import { CommonService } from '@common/common.service';
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { ApiDocs } from '@common/variable/dsl';
+import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { isNil } from '@util/isNil';
 import crypto from 'crypto';
@@ -65,7 +66,7 @@ export class ProfilesService {
         },
       };
     } catch (error) {
-      throw new NotFoundException('profile');
+      throw new ApiDocs.DslNotFoundProfile();
     }
   }
 
@@ -75,7 +76,7 @@ export class ProfilesService {
     });
 
     if (isNil(profile)) {
-      throw new NotFoundException('profile');
+      throw new ApiDocs.DslNotFoundProfile();
     }
 
     const originalname = file.originalname;
@@ -102,7 +103,15 @@ export class ProfilesService {
     );
   }
 
-  remove(userId: number) {
+  async remove(userId: number) {
+    const profile = await this.profileRepository.findOne({
+      where: { userId },
+    });
+
+    if (isNil(profile)) {
+      throw new ApiDocs.DslNotFoundProfile();
+    }
+
     return this.profileRepository.delete({ userId });
   }
 }
